+++ TMSystem/src/java/com/corebits/ericsson/tms/models/LoanApplication.java
@@ -55,11 +55,11 @@ public class LoanApplication implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "loan_amount")
-    private BigDecimal loanAmount;
+    private double loanAmount;
     @Basic(optional = false)
     @NotNull
     @Column(name = "annual_interest_rate")
-    private BigDecimal annualInterestRate;
+    private double annualInterestRate;
     @Basic(optional = false)
     @NotNull
     @Column(name = "loan_start_date")
@@ -68,7 +68,7 @@ public class LoanApplication implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "monthly_payment_amount")
-    private BigDecimal monthlyPaymentAmount;
+    private double monthlyPaymentAmount;
     @Basic(optional = false)
     @NotNull
     @Column(name = "number_of_payment")
@@ -76,11 +76,11 @@ public class LoanApplication implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "total_interest")
-    private BigDecimal totalInterest;
+    private double totalInterest;
     @Basic(optional = false)
     @NotNull
     @Column(name = "total_cost_of_loan")
-    private BigDecimal totalCostOfLoan;
+    private double totalCostOfLoan;
     @Basic(optional = false)
     @NotNull
     @Column(name = "date_of_application")
@@ -107,7 +107,7 @@ public class LoanApplication implements Serializable {
         this.id = id;
     }
 
-    public LoanApplication(Integer id, StaffMember memberId, BigDecimal loanAmount, BigDecimal annualInterestRate, Date loanStartDate, BigDecimal monthlyPaymentAmount, int numberOfPayment, BigDecimal totalInterest, BigDecimal totalCostOfLoan, BigDecimal monthlyPrincipal, Date dateOfApplication) {
+    public LoanApplication(Integer id, StaffMember memberId, double loanAmount, double annualInterestRate, Date loanStartDate, double monthlyPaymentAmount, int numberOfPayment, double totalInterest, double totalCostOfLoan, double monthlyPrincipal, Date dateOfApplication) {
         this.id = id;
         this.memberId = memberId;
         this.loanAmount = loanAmount;
@@ -128,19 +128,19 @@ public class LoanApplication implements Serializable {
         this.id = id;
     }  
 
-    public BigDecimal getLoanAmount() {
+    public double getLoanAmount() {
         return loanAmount;
     }
 
-    public void setLoanAmount(BigDecimal loanAmount) {
+    public void setLoanAmount(double loanAmount) {
         this.loanAmount = loanAmount;
     }
 
-    public BigDecimal getAnnualInterestRate() {
+    public double getAnnualInterestRate() {
         return annualInterestRate;
     }
 
-    public void setAnnualInterestRate(BigDecimal annualInterestRate) {
+    public void setAnnualInterestRate(double annualInterestRate) {
         this.annualInterestRate = annualInterestRate;
     }
 
@@ -152,11 +152,11 @@ public class LoanApplication implements Serializable {
         this.loanStartDate = loanStartDate;
     }
 
-    public BigDecimal getMonthlyPaymentAmount() {
+    public double getMonthlyPaymentAmount() {
         return monthlyPaymentAmount;
     }
 
-    public void setMonthlyPaymentAmount(BigDecimal monthlyPaymentAmount) {
+    public void setMonthlyPaymentAmount(double monthlyPaymentAmount) {
         this.monthlyPaymentAmount = monthlyPaymentAmount;
     }
 
@@ -168,19 +168,19 @@ public class LoanApplication implements Serializable {
         this.numberOfPayment = numberOfPayment;
     }
 
-    public BigDecimal getTotalInterest() {
+    public double getTotalInterest() {
         return totalInterest;
     }
 
-    public void setTotalInterest(BigDecimal totalInterest) {
+    public void setTotalInterest(double totalInterest) {
         this.totalInterest = totalInterest;
     }
 
-    public BigDecimal getTotalCostOfLoan() {
+    public double getTotalCostOfLoan() {
         return totalCostOfLoan;
     }
 
-    public void setTotalCostOfLoan(BigDecimal totalCostOfLoan) {
+    public void setTotalCostOfLoan(double totalCostOfLoan) {
         this.totalCostOfLoan = totalCostOfLoan;
     }
